Sasaiste ar calendar shandule un acttivities

+++ src/main/java/lv/venta/services/impl/CalendarScheduleService.java
@@ -12,48 +12,25 @@ import lv.venta.repos.ICalendarScheduleRepo;
 import lv.venta.repos.IStudioProgrammRepo;
 import lv.venta.services.ICalendarScheduleService;
 
-import java.time.Year;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class CalendarScheduleService implements ICalendarScheduleService{
 
 	private ICalendarScheduleRepo calendarScheduleRepo;
 	private IStudioProgrammRepo studioProgrammRepo;
-	
 	private ICalendarRepo calendarRepo;
+	
 	
 	@Autowired
 	public CalendarScheduleService(ICalendarScheduleRepo calendarScheduleRepo,IStudioProgrammRepo studioProgrammRepo, ICalendarRepo calendarRepo) {
 	    this.calendarScheduleRepo = calendarScheduleRepo;
 	    this.studioProgrammRepo = studioProgrammRepo;
-	    
 	    this.calendarRepo = calendarRepo;
 	}
-	
-	@Override
-    public List<CalendarScheduleDTO> getAllCalendarSchedules() {
-        List<CalendarSchedule> calendarSchedules = (List<CalendarSchedule>) calendarScheduleRepo.findAll();
 
-        // Saraksts kur glabāsies DTO objekti
-        List<CalendarScheduleDTO> calendarScheduleDTOList = new ArrayList<>();
-
-        // Loops caur  CalendarSchedule un parveidot dto objektos
-        for (CalendarSchedule calendarSchedule : calendarSchedules) {
-            CalendarScheduleDTO dto = new CalendarScheduleDTO();
-            dto.setGads(calendarSchedule.getGads());
-            dto.setStudioProgrammTitle(calendarSchedule.getStudioProgramm().getTitle());
-            dto.setActivity(calendarSchedule.getActivities().get(0).getActivity());
-            dto.setActivityEndDate(calendarSchedule.getActivities().get(0).getActivityEndDate());
-            dto.setActivityImplementation(calendarSchedule.getActivities().get(0).getActivityImplementation());
-
-            // pievienot dto sarakstam
-            calendarScheduleDTOList.add(dto);
-        }
-
-        return calendarScheduleDTOList;
-    }
 	
 	public void addCalendarSchedule(CalendarScheduleDTO calendarScheduleDTO) {
         // 1.Jaatrod eksistejosa studiju programma pec nosaukuma
@@ -81,25 +58,44 @@ public class CalendarScheduleService implements ICalendarScheduleService{
         }
         
     }
-	
+
+    	
 	@Override
 	public void removeCalendarSchedule(long idCalendar) {
-	    CalendarSchedule calendarSchedule = calendarScheduleRepo.findById(idCalendar)
-	            .orElseThrow(() -> new IllegalArgumentException("Kalendāra grafiks ar norādīto ID nav atrasts"));
+	    Optional<CalendarSchedule> optionalCalendarSchedule = calendarScheduleRepo.findById(idCalendar);
 
-	    List<CalendarActivity> activities = calendarSchedule.getActivities();
-	    
-	    for (CalendarActivity activity : calendarSchedule.getActivities()) {
-	    	calendarRepo.delete(activity);
+	    if (optionalCalendarSchedule.isPresent()) {
+	        CalendarSchedule calendarSchedule = optionalCalendarSchedule.get();
+
+	        calendarScheduleRepo.delete(calendarSchedule);
+	    } else {
+	        throw new IllegalArgumentException("Kalendāra ieraksts ar ID " + idCalendar + " nav atrasts.");
 	    }
-
-	    calendarScheduleRepo.delete(calendarSchedule);
 	}
-
+	
+	
 	@Override
-	public List<CalendarSchedule> getCalendarSchedules() {
-	    return (List<CalendarSchedule>) calendarScheduleRepo.findAll();
-	}
+    public List<CalendarScheduleDTO> getAllCalendarSchedules() {
+        List<CalendarSchedule> calendarSchedules = (List<CalendarSchedule>) calendarScheduleRepo.findAll();
 
+        // Saraksts kur glabāsies DTO objekti
+        List<CalendarScheduleDTO> calendarScheduleDTOList = new ArrayList<>();
+
+        // Loops caur  CalendarSchedule un parveidot dto objektos
+        for (CalendarSchedule calendarSchedule : calendarSchedules) {
+            CalendarScheduleDTO dto = new CalendarScheduleDTO();
+            dto.setId(calendarSchedule.getIdCalendar());
+            dto.setGads(calendarSchedule.getGads());
+            dto.setStudioProgrammTitle(calendarSchedule.getStudioProgramm().getTitle());
+            dto.setActivity(calendarSchedule.getActivities().get(0).getActivity());
+            dto.setActivityEndDate(calendarSchedule.getActivities().get(0).getActivityEndDate());
+            dto.setActivityImplementation(calendarSchedule.getActivities().get(0).getActivityImplementation());
+
+            // pievienot dto sarakstam
+            calendarScheduleDTOList.add(dto);
+        }
+
+        return calendarScheduleDTOList;
+    }
 	
 }
